Avoid storing rangetype oids within multiranges
This saves a few bytes for each range within a multirange. Instead of
storing RangeTypes within a multirange, we store ShortRangeTypes which
exclude the range type oid. That means we can move the range flags byte
from after the range bounds to where the Oid used to live. Because
ranges and their bounds must be aligned, that should save 4 bytes per
range.

Because almost every multirange function and operation uses the internal
ranges to compute a result, we usually need a way to convert
ShortRangeTypes back into RangeTypes. For that we use the EXPANDED TOAST
infrastructure.