WIP: FOR PORTION OF trigger variables
I thought I should expose the tg_temporal struct that we give to
triggers where there is a FOR PORTION OF UPDATE/DELETE to plpgsql. The
problem is that tg_targetRange is an anyrange, and I don't see a way to
make that work as a plpgsql variable. The actual type could even be
different for different calls of the same trigger function, if there is
more than one range column/PERIOD. So I think I'm going to just save a
commit of this to remind me where things are and why I don't think I can
implement it.