fixup fks cascade: Fix possible overflow in temporal FK queryoids
This arrays holds the type oids for all the parameters we use in the SPI
queries implementing foreign keys. When a temporal foreign key has
CASCADE/SET NULL/SET DEFAULT, we give a last parameter with the targeted
range, then we reference it in FOR PORTION OF FROM/TO. So we need an
extra spot in the array or you could overflow it.