fixup periods: use full index in get_relation_period_oid
For some reason Vik's original code was using only the relid part of the
index, not the period name part. And then it was making sure there were
no duplicates. But it's a UNIQUE index. So let's just take advantage of
that (the same way get_relation_constraint_oid does).