fixup periods: Exclude GENERATED column from CREATE TABLE LIKE
When you use LIKE inside CREATE TABLE to copy columns, we should skip
the GENERATED columns that implement PERIODs. Because they don't make
sense without their PERIODs, and LIKE never copies PERIODs. In fact it
causes problems because then if you use LIKE and also try to add a
PERIOD of the same name as in the parent table, you get a conflict: not
between the old & new period, but between the old period's column and
the new period.