fixup pks: fix implicit NOT NULL constraint on PK WITHOUT OVERLAPS column
This was out-of-date after Alvaro's named NOT NULL constraint commit, so
that it caused ALTER TABLE to fail when trying to add a temporal
PK/UNIQUE constraint.

Also it turns out we were forcing the WITHOUT OVERLAPS column to NOT
NULL when it was just a UNIQUE constraint, which isn't really right. It
should only happens for PKs.