fixup fpo: copy FPO struct to each ResultRelInfo
This fixes a segfault with partitioned tables when there is a FOR
PORTION OF cross-partition update. That is implemented as a DELETE on
the old partition then an INSERT on the new one. But we were only
storing the FPO struct on the first ResultRelInfo, so the delete could
find it NULL (if it wasn't the first partition).

We can actually fix that by always using
mtstate->resultRelInfo->ri_forPortionOf in ExecForPortionOfLeftovers,
but that doesn't solve triggers, which also need access to the struct
and only get their own ResultRelInfo, not the whole ModifyTableContext.

It's too bad, because if there are a lot of partitions, we copy the FPO
struct to every one, whether they do anything in the query or not.