fixup fks: fix jian he NO ACTION bug
The foreign key should forbid an UPDATE/DELETE but doesn't.

This is because we call ri_Check_Pk_Match first, which see if the PK
table has a (new) row that matches the keys from the row we're deleting.
Since the FK comparison operator for the range column is &&, it detects
the inserted "leftovers" rows and thinks it doesn't need to do the FK
check. The fix is to use a similar aggregation to our temporal FK check
to see if the new row(s) *completely cover* the lost row.

This commit has the fix but no test code because you can't reproduce the
problem without FOR PORTION OF. So I'll add test code there.

I wonder if this optimization actually helps in the temporal case. Even
in the non-temporal case, it means running two SQL checks instead of
one.