Fix jian he's RESTRICT bug
The RESTRICT FK constraint could fail when it shouldn't if used with FOR
PORTION OF. This is because it calls ri_restrict which calls
ri_PerformCheck and that needs to know what range was targeted (since it
may not be the whole row).

Includes new tests to show the bug.

Without this fix we would also get errors (later) in ON UPDATE/DELETE
SET DEFAULT, since that can also call ri_restrict sometimes.