fixup fpo: remove UNBOUNDED from FOR PORTION OF grammar
We wanted to allow people to say `FOR PORTION OF valid_at FROM UNBOUNDED
TO UNBOUNDED` and translate those to NULLs, because it's a little weird
to say `FROM NULL TO NULL`. But jian he pointed out I can't just look
for these as ColumnDefs, because then we also allow `FROM "UNBOUNDED" TO
"UNBOUNDED"` and still turn them into nulls, which seems wrong: the
quotes should allow those to be *actual column refs* (even though those
are illegal in this context).

So I tried putting `FROM UNBOUNDED` and `TO UNBOUNDED` in gram.y, which
is the right way to do it. But that is causing a reduce/reduce conflict.
There are comments in gram.y about how `UNBOUNDED` is already a mess in
the SQL spec (not a reserved keyword yet used in window function frames,
which causes ambiguity). I also tried going back to MINVALUE and MAXVALUE
(my original idea), but I was still getting conflicts there too. I'm just
taking out the UNBOUNDED stuff for now, but maybe I can make
MINVALUE/MAXVALUE work eventually.