fixup fpo: use operator translation for FPO
We need overlaps (&&) and also intersect (*). There is no stratnum for
intersect, so this adds it. But it's neither a search ('s') nor
ordering ('o') operator, so I added a new oprpurpose, "other" ('x').

Maybe this is all wrong since we're not really using it as an *index*
operator. Could we add something to pg_operator instead, like
oprcanintersect? But then I'll need oprcanleftovers too.

I should probably make FPO require that you have a GiST opclass defined
for the type of the valid_at column. These operations are *sort of*
index things because you probably have a PK and we want to be consistent
with it.