fixup pks: remove range type restriction
Took out the check in Analysis Phase that the WITHOUT OVERLAPS column
was a range type. Now the error message is not as nice (about gist
opclasses) but OTOH it gives you more of a clue how to support the type
you want. (But if you just mistyped the column type then it's harder.)

Also added a basic test for creating a multirange PK.