Add without_portion GiST support proc
This new support proc is used by RESTRICT foreign keys to compute the
portion of history that was lost when the application-time bounds of a
record change. This commit defines implementations for ranges and
multiranges.

The procs return SETOF their input type and work like minus but don't
fail on splits. The results never contain empty elements.

When we implement FOR PORTION OF, we will use these procs simiarly: to
compute leftovers that weren't touched by the UPDATE/DELETE.

Author: Paul Jungwirth