fixup fpo: support non-range column types
Still needs tests, but now we will call the @- oper (or rather its proc)
so that you can use multiranges or anything else.

This adds @- to pg_amop also, which has the same weirdness as putting
intersect (*) there---i.e. it is a 'p' (portion) purpose, not 's'
(search) or 'o' (ordering).