fixup fpo: fix FOR PORTION OF with partitions
There were lots of problems using the wrong attnos if they differed
between partitions. This commit generates a separate TupleTableSlot for
each one and does the necessary conversions to (1) read the old row with
the leaf partition's tuple descriptor (2) write the leftovers with the
root partition's tuple descriptor. And we initialize tuple routing so
that when we use ExecInsert for the leftovers, they land in the right
place.