fixup fpo: remove TODO about tuple locking
jian he feedback 8 Jan 2024

So I know that when EvalPlanQual *fails* we lock the tuple before coming
here, but what if it succeeds? Does table_tuple_update lock the tuple?
Tracking that down is hard because this follows a function pointer for
the AM.

But in access/tableam.h I see that a TM_Result of TM_Ok means the tuple
has been locked. We won't get to Update/DeleteEpilogue until that
happens, so we are safe.