fixup periods: change dependency type to INTERNAL
INTERNAL is the correct dependency type, but findDependentObjects
doesn't do the right thing when the internal object is a column (i.e.
something with a subobjectid). It thinks you are dropping the whole
table and objects. This commit fixes that, which means we can remove the
hack about using a NORMAL dependency one way and an AUTO dependency the
other way.