fixup periods: fix error message changing the type of a start/end column
If you change the type of a start/end column used by a PERIOD, it should
fail. It was, but the error message could complain about the invisible
GENERATED column instead (non-deterministically, based on which
pg_depend record was seen first). This commit sees that the GENERATED
column is used by a PERIOD and skips it, so that we always give an error
message about the PERIOD instead.