fixup fks: fix range-only code
We still had ri_RangeAttributeNeedsCheck which doesn't support
multiranges or other types. Instead we need to save the "contained-by"
operator (not the once with assymetic types but the normal one), and
then use that. And in fact we can do that with just ri_AttributesEqual.

I also added lots of multirange FK tests, since evidentally those are
needed! (Still no multirange tests with FOR PORTION OF here though,
since that comes later.)