Forbid empty ranges/multiranges in WITHOUT OVERLAPS columns
Because 'empty' && 'empty' is false, the temporal PK/UQ constraint allows
duplicates, which is confusing to users and breaks internal
expectations. For instance when GROUP BY checks functional dependencies
on the PK, it allows selecting other columns from the table, but in the
presence of duplicate keys you could get the value from any of their
rows. So we need to forbid empties.

Added pg_class.relperiods to track the number of WITHOUT OVERLAPS index
constraints on the table (similar to relchecks).

Increment relperiods when a WITHOUT OVERLAPS constraint is added, and
decrement it when removed.

In the relcache, if relperiods > 0 then get the attnos of all columns
used as the WITHOUT OVERLAPS part of a constraint.

Check for empty values in WITHOUT OVERLAPS columns when recording an
inserted/updated row.

Check for already-existing empties when adding a new temporal PK/UQ.

This all means we can only support ranges and multiranges for temporal
PK/UQs. So I added a check and updated the docs and tests.