Forbid empty ranges/multiranges in WITHOUT OVERLAPS columns
Because 'empty' && 'empty' is false, the temporal PK/UQ constraint allows
duplicates, which is confusing to users and breaks internal
expectations. For instance when GROUP BY checks functional dependencies
on the PK, it allows selecting other columns from the table, but in the
presence of duplicate keys you could get the value from any of their
rows. So we need to forbid empties.

This all means we can only support ranges and multiranges for temporal
PK/UQs. So I added a check and updated the docs and tests.