Improve comment about snapshot macros
The comment mistakenly had "the others" for "the other", but this
commit also reorders the comment so it matches the macros below. Now we
describe the levels in increasing strictness. Finally, it seems easier
to follow if we introduce one level at a time, rather than describing
two, followed by "the other" (and then jumping back to one of the first
two).