Add docs section for temporal tables, with primary keys
This section introduces temporal tables, with a focus on Application Time (which
we support) and only a brief mention of System Time (which we don't). It covers
temporal primary keys and unique constraints. Temporal foreign keys are
documented in the next commit. We will document temporal update/delete and
periods as we add those features.

This commit also adds glossary entries for temporal table, application
time, and system time.

Author: Paul A. Jungwirth <pj@illuminatedcomputing.com>