Add docs chapter for temporal tables
This commit tries to outline the complete functionality described by the
SQL:2011 standard, which we hope to achieve. So it includes sections for
Application Time and System Time, but it notes that System Time is not
yet implemented. Likewise it covers temporal primary keys and unique
constraints in detail, but it only notes that temporal update/delete are
not yet supported.

Temporal foreign keys and periods are documented in the next commits.

This commit also adds glossary entries for temporal table, application
time, and system time.