Add range_get_constructor2
Look up the two-arg constructor for a given rangetype. We need this for
UPDATE/DELETE FOR PORTION OF, so that we can build a range from the FROM/TO
bounds.

There doesn't seem to be an easy way to find the constructor. The rule is that
the function has the same name as the rangetype, with arguments matching the
range's subtype. Ideally we could just use the range's type oid, but I see no
way to do that. Using pg_depend doesn't help: there are no entries for built-in
rangetypes, only user-defined ones.

Author: Paul A. Jungwirth <pj@illuminatedcomputing.com>