wip: domains: validating leftovers works but is ugly
I added an isDomain bool to ForPortionOfExpr so I can skip the work if it's not
needed. I'm passing a NULL extra, so the domain code has to do extra work. I
could put a pointer in the FPO state struct to avoid that. But also I wonder if
I should build an expr node using coerce_type and use that? Then the
without_portion func just needs to return something coercible to the column, and
domain checks should happen automatically (I think).