gist tests pass, without_overlaps is broken
We detect duplicates and fail for CREATE INDEX and INSERT. We don't do any
concurrency work yet (returning xwait and waiting on it).

Also our changes broke the WITHOUT OVERLAPS tests. The index sees that it's a
"unique" index, and it fails. But the records are actually unique. I bet it's
using overlaps instead of equals for the excludeFn.