Fix problems, especially with multiranges
We need to use the original heap Datums when comparing rows, because if the
opclass has a compress function, the index Datum is not the same. In particular
this happens with multiranges.

We are going to the heap anyway for visibility info, so it's not any extra cost
to do it properly, and we don't need to bother with optimizations when we have
no compress function.

Checking heap Datums does mean we need to convert the index attnum to the heap
attnum!

For the new tuple, we already have the original heap Datums just a couple
function calls up, so I pass those down to avoid an extra lookup. Note that here
we should use the index attnum to index into the array.

I also fixed some issues with building the error message. It also needs to use
the non-compressed Datums, since it passes them to the type's out function.